extract record blocking
